Run the desire specification from a SpANOVA object

This function runs the desire specification from a SpANOVA object, using the data prepare from the original specification

Usage

inla.rerun.SpANOVA(
  SpANOVA_obj,
  n_mod,
  gr,
  scale.mod = TRUE,
  sp.prior = "sdunif",
  pc.prec.val = c(1, 0.01),
  sp.copy.fixed = TRUE,
  verbose.INLA = FALSE
)

Arguments

SpANOVA_obj

Object generated after running the inla.SpANOVA.2x2 function

n_mod

Number of the model that you want to run, using the value reflected in the SpANOVA summary table

gr

Graph for the underlying spatial structure

scale.mod

Scale copies of random spatial effects or not, default is TRUE

sp.prior

Select prior for the random spatial effect, options are sdunif and pc.prec

pc.prec.val

Define values por the pc prior in case it was chosen. Default values are c(1, 0.01)

sp.copy.fixed

Fix copied values for the random spatial effects, default is TRUE

verbose.INLA

Change verbose value for INLA in case you want to follow the approximation

Value

Object generated by INLA, including groups and sp_effects data from the SpANOVA specification